requirements.txt
absl-py==0.9.0
alembic==1.4.2
astunparse==1.6.3
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
DateTime==4.3
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-MySQL==1.5.1
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gast==0.3.3
get==2019.4.13
google-api-core==1.21.0
google-api-python-client==1.10.0
google-auth==1.19.2
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.30.0
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mysql-connector-python==8.0.21
mysqlclient==1.4.6
numpy==1.19.0
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.5
Pillow==7.2.0
post==2019.4.13
protobuf==3.12.2
public==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==0.9.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
query-string==0.0.0
request==2019.4.13
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.5.1
selenium==3.141.0
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
termcolor==1.1.0
threadpoolctl==2.1.0
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
zope.interface==5.1.0
